pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "testing-pipeline-llm-chatbot"
version = "0.1.0"
description = "Knowledge Sharing Demo: AI Testing Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.0.1",
    "litellm>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "testcontainers>=4.0.0",
    "ruff>=0.3.0",
    "ragas>=0.1.0"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
